Added  Azure-ReadOCR-PDFScript.ps1
Added script that scans for PDFs in a folder and automatically submits them to Azure Document Intelligence to be OCR'd by the Read Model and outputs both a searchable PDF and a JSON file. Requires you use an API Key from Azure as these models aren't free. Your Azure tenant this ties to will need proper stuff installed for Azure Document Intelligence to function. This changes constantly so you'll have to research and set this up yourself.